Rename filterterm module to filter-term
Adapt for our defined naming scheme.
