+++ src/web/components/dashboard/display/cvss/CvssDisplay.jsx
@@ -5,7 +5,7 @@
 
 import React from 'react';
 import Filter from 'gmp/models/filter';
-import FilterTerm from 'gmp/models/filter/filterterm';
+import FilterTerm from 'gmp/models/filter/filter-term';
 import {isDefined} from 'gmp/utils/identity';
 import BarChart from 'web/components/chart/Bar';
 import transformCvssData from 'web/components/dashboard/display/cvss/cvss-transform';
