+++ src/web/pages/notes/dashboard/ActiveDaysDisplay.jsx
@@ -6,7 +6,7 @@
 import React from 'react';
 import {_, _l} from 'gmp/locale/lang';
 import Filter, {NOTES_FILTER_FILTER} from 'gmp/models/filter';
-import FilterTerm from 'gmp/models/filter/filterterm';
+import FilterTerm from 'gmp/models/filter/filter-term';
 import {
   NOTE_ACTIVE_UNLIMITED_VALUE,
   NOTE_INACTIVE_VALUE,
